+++ nahsor/client.py
@@ -1,4 +1,9 @@
 # -*- coding:utf-8 -*-
+'''
+作者：浪晋
+时间：2018-6-2
+说明：启动接口测试的入口
+'''
 from app.core.testutils import run_http_test, insert_test_result
 from app.core.testutils import extract_global_values
 from app.core.testutils import get_global_values, get_validata_test
