+++ nahsor/output.py
@@ -1,0 +1,6 @@
+# -*- coding:utf-8 -*-
+'''
+作者：浪晋
+时间：2018-6-2
+说明：控制日志输出的格式
+'''
