+++ requirements.txt
@@ -1,4 +1,2 @@
-flask==0.12.2
-pymysql==0.8.0
 requests==2.18.4
-flask-cors==3.0.4
+coloredlogs==10.0
